+++ samples/Sentry.Samples.Maui/MauiProgram.cs
@@ -8,6 +8,8 @@ public static class MauiProgram
             .UseSentry(options =>
             {
                 options.Dsn = "https://eb18e953812b41c3aeb042e666fd3b5c@o447951.ingest.sentry.io/5428537";
+                options.Debug = true;
+                options.MaxBreadcrumbs = int.MaxValue; // TODO: reduce breadcrumbs, remove this
             })
             .ConfigureFonts(fonts =>
             {